requirements.txt
requests==2.28.0
allure-pytest==2.9.45
pytest==7.1.2
email-validator==1.2.1